requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
faker
python-dotenv
Flask-Paginate
werkzeug
SQLAlchemy
urllib3